requirements.txt
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.1
Flask==2.3.2
google-api-core==2.11.0
google-auth==2.18.1
google-cloud-appengine-logging==1.3.0
google-cloud-audit-log==0.2.5
google-cloud-core==2.3.2
google-cloud-logging==3.5.0
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
greenlet==2.0.1
grpc-google-iam-v1==0.12.6
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
Pillow==9.5.0
playwright==1.33.0
pluggy==1.0.0
proto-plus==1.22.2
protobuf==4.23.1
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyee==9.0.4
pytest==7.3.1
pytest-base-url==2.0.0
pytest-html==3.2.0
pytest-metadata==2.0.4
pytest-playwright==0.3.3
python-slugify==8.0.1
requests==2.30.0
rsa==4.9
six==1.16.0
text-unidecode==1.3
tomli==2.0.1
typing-extensions==4.5.0
urllib3==1.26.15
uuid==1.30
Werkzeug==2.3.4
zipp==3.15.0
